create events folder, move onComplete in
