+++ core/_interfaces/index.ts
@@ -1,12 +1,9 @@
 import { ITestCase } from "./test-case.i";
 import { ITestFixture } from "./test-fixture.i";
 import { ITest } from "./test.i";
-import { ITestCompleteEvent, IOnTestCompleteCBFunction } from "./test-complete.i";
 
 export {
     ITestCase,
     ITestFixture,
-    ITest,
-    ITestCompleteEvent,
-    IOnTestCompleteCBFunction
+    ITest
 };
