+++ core/events/index.ts
@@ -1,0 +1,6 @@
+import { ITestCompleteEvent, IOnTestCompleteCBFunction } from "./test-complete.i";
+
+export {
+    ITestCompleteEvent,
+    IOnTestCompleteCBFunction
+};
